[Cli] Move middleware from Routing to Server.
