+++ src/Valkyrja/Cli/Routing/Provider/ServiceProvider.php
@@ -31,7 +31,6 @@ use Valkyrja\Cli\Routing\Constant\AllowedClasses;
 use Valkyrja\Cli\Routing\Data\Data;
 use Valkyrja\Cli\Routing\Dispatcher\Contract\RouterContract;
 use Valkyrja\Cli\Routing\Dispatcher\Router;
-use Valkyrja\Cli\Routing\Middleware\RouteNotMatched\CheckCommandForTypoMiddleware;
 use Valkyrja\Container\Manager\Contract\ContainerContract;
 use Valkyrja\Container\Provider\Provider;
 use Valkyrja\Dispatch\Dispatcher\Contract\DispatcherContract;
@@ -46,10 +45,9 @@ final class ServiceProvider extends Provider
     public static function publishers(): array
     {
         return [
-            CollectorContract::class             => [self::class, 'publishAttributeCollector'],
-            RouterContract::class                => [self::class, 'publishRouter'],
-            CollectionContract::class            => [self::class, 'publishCollection'],
-            CheckCommandForTypoMiddleware::class => [self::class, 'publishCheckCommandForTypoMiddleware'],
+            CollectorContract::class  => [self::class, 'publishAttributeCollector'],
+            RouterContract::class     => [self::class, 'publishRouter'],
+            CollectionContract::class => [self::class, 'publishCollection'],
         ];
     }
 
@@ -63,7 +61,6 @@ final class ServiceProvider extends Provider
             CollectorContract::class,
             RouterContract::class,
             CollectionContract::class,
-            CheckCommandForTypoMiddleware::class,
         ];
     }
 
@@ -143,19 +140,5 @@ final class ServiceProvider extends Provider
                 ...$collector->getRoutes(...$controllers)
             );
         }
-    }
-
-    /**
-     * Publish the check command for typo middleware service.
-     */
-    public static function publishCheckCommandForTypoMiddleware(ContainerContract $container): void
-    {
-        $container->setSingleton(
-            CheckCommandForTypoMiddleware::class,
-            new CheckCommandForTypoMiddleware(
-                router: $container->getSingleton(RouterContract::class),
-                collection: $container->getSingleton(CollectionContract::class),
-            )
-        );
     }
 }
